majobacore/requirements/production.txt
# Production requirements
-r base.txt

sentry-sdk>=1.35.0
django-storages>=1.14.0
boto3>=1.29.0
newrelic>=9.2.0
